Rename filelock package to pidlock
